2-й ЛК

+++ app/views/projects-administrator/clients/show.blade.php
@@ -30,7 +30,6 @@
             <tbody>
             <?php $tasks_total_time = 0;?>
             <?php $earnMoneyCurrentDate = costCalculation(NULL,['tasks' => $tasks]);?>
-            <?php $earnMoneyTotal = costCalculation();?>
             @foreach($tasks as $task)
                 <?php $tasks_total_time += (getLeadTimeMinutes($task)+floor($task->lead_time/60));?>
                 <tr {{ ($task->start_status && !$task->stop_status) ? 'class="success"' : '' }}>
@@ -42,14 +41,10 @@
                         @endif
                     </td>
                     <td>
-                        {{ culcLeadTime($task) }} / {{ isset($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings'],2,'.',' ').' руб.' : '' }}
-                        @if(isset($earnMoneyTotal[$task->project->id][$task->id]['overdose']) && $earnMoneyTotal[$task->project->id][$task->id]['overdose'] == 1)
-                            <br><span class="label label-danger">Превышен допустимый лимит бюджета</span>
-                            <br><span class="label label-info">Текущий заработок: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['overdose_money'],2,'.',' ').' руб.' }}</span>
-                            <br><span class="label label-info">Доступный бюджет: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['budget'],2,'.',' ').' руб.' }}</span>
-                        @endif
+                        {{ culcLeadTime($task) }} / {{ isset($earnMoneyCurrentDate[$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->id]['earnings'],2,'.',' ').' руб.' : '' }}
                     </td>
                     <td>
+                    @if($task->user_id == Auth::user()->id)
                         {{ Form::open(array('route'=>array('timesheets.run_timer'),'method'=>'POST','style'=>'display:inline-block')) }}
                         {{ Form::hidden('task',$task->id) }}
                         @if($task->start_status && !$task->stop_status)
@@ -60,12 +55,19 @@
                             {{ Form::submit('Начать',['class'=>'btn btn-default']) }}
                         @endif
                         {{ Form::close() }}
+                    @endif
                     </td>
-                    <td><a href="{{ URL::route('timesheets.edit',$task->id) }}" class="btn btn-success">Редактировать</td>
                     <td>
+                    @if($task->user_id == Auth::user()->id)
+                        <a href="{{ URL::route('timesheets.edit',$task->id) }}" class="btn btn-success">Редактировать</a>
+                    @endif
+                    </td>
+                    <td>
+                    @if($task->user_id == Auth::user()->id)
                         {{ Form::open(array('route'=>array('timesheets.destroy',$task->id),'method'=>'DELETE','style'=>'display:inline-block')) }}
                         {{ Form::submit('Удалить',['class'=>'btn btn-danger']) }}
                         {{ Form::close() }}
+                    @endif
                     </td>
                 </tr>
             @endforeach
