+++ app/views/projects-administrator/projects/show.blade.php
@@ -5,17 +5,31 @@
     <div class="jumbotron" style="background-image: url(/uploads/images/1422287024_1908.jpg)">
         <h1>{{ $project->title }}</h1>
         <p class="lead">{{ $project->description }}</p>
+    @if($access)
         @if(!empty($project->client))
             <a href="{{ URL::route('clients.show',$project->client->id) }}" class=""><h5>{{ $project->client->title }}</h5></a>
         @endif
+    @else:
+        @if(!empty($project->client))<h5>{{ $project->client->title }}</h5>@endif
+    @endif
+    @if($access)
         <a role="button" href="{{ URL::route('projects.edit',$project->id) }}" class="btn btn-primary btn-sm">Редактировать</a>
         {{ Form::open(array('route'=>array('projects.destroy',$project->id),'method'=>'DELETE','style'=>'display:inline-block')) }}
             {{ Form::submit('Удалить',['class'=>'btn btn-danger btn-sm']) }}
         {{ Form::close() }}
+    @endif
     </div>
     @if($project->team->count())
     <div class="container marketing">
         <div class="row">
+        @if($project->superior->count())
+            <div class="col-lg-4">
+                <img class="img-circle" data-src="holder.js/140x140/auto/sky" alt="">
+                <h2>{{ getInitials($project->superior->fio) }}</h2>
+                <p>{{ $project->superior->position }}</p>
+                <p><a class="btn btn-default" href="{{ URL::route('cooperators.show',$project->superior->id) }}" role="button">Подробнее &raquo;</a></p>
+            </div>
+        @endif
         @foreach($project->team as $user)
             <div class="col-lg-4">
                 <img class="img-circle" data-src="holder.js/140x140/auto/sky" alt="">
@@ -34,7 +48,6 @@
             <tbody>
             <?php $tasks_total_time = 0;?>
             <?php $earnMoneyCurrentDate = costCalculation(NULL,['tasks' => $tasks]);?>
-            <?php $earnMoneyTotal = costCalculation();?>
             @foreach($tasks as $task)
                 <?php $tasks_total_time += (getLeadTimeMinutes($task)+floor($task->lead_time/60));?>
                 <tr {{ ($task->start_status && !$task->stop_status) ? 'class="success"' : '' }}>
@@ -46,14 +59,10 @@
                         @endif
                     </td>
                     <td>
-                        {{ culcLeadTime($task) }} / {{ isset($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings'],2,'.',' ').' руб.' : '' }}
-                        @if(isset($earnMoneyTotal[$task->project->id][$task->id]['overdose']) && $earnMoneyTotal[$task->project->id][$task->id]['overdose'] == 1)
-                            <br><span class="label label-danger">Превышен допустимый лимит бюджета</span>
-                            <br><span class="label label-info">Текущий заработок: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['overdose_money'],2,'.',' ').' руб.' }}</span>
-                            <br><span class="label label-info">Доступный бюджет: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['budget'],2,'.',' ').' руб.' }}</span>
-                        @endif
+                        {{ culcLeadTime($task) }} @if($task->user_id == Auth::user()->id)/ {{ isset($earnMoneyCurrentDate[$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->id]['earnings'],2,'.',' ').' руб.' : '' }}@endif
                     </td>
                     <td>
+                    @if($task->user_id == Auth::user()->id)
                         {{ Form::open(array('route'=>array('timesheets.run_timer'),'method'=>'POST','style'=>'display:inline-block')) }}
                         {{ Form::hidden('task',$task->id) }}
                         @if($task->start_status && !$task->stop_status)
@@ -64,12 +73,18 @@
                             {{ Form::submit('Начать',['class'=>'btn btn-default']) }}
                         @endif
                         {{ Form::close() }}
+                    @endif
                     </td>
-                    <td><a href="{{ URL::route('timesheets.edit',$task->id) }}" class="btn btn-success">Редактировать</td>
                     <td>
+                    @if($task->user_id == Auth::user()->id)
+                        <a href="{{ URL::route('timesheets.edit',$task->id) }}" class="btn btn-success">Редактировать</a></td>
+                    @endif
+                    <td>
+                    @if($task->user_id == Auth::user()->id)
                         {{ Form::open(array('route'=>array('timesheets.destroy',$task->id),'method'=>'DELETE','style'=>'display:inline-block')) }}
                         {{ Form::submit('Удалить',['class'=>'btn btn-danger']) }}
                         {{ Form::close() }}
+                    @endif
                     </td>
                 </tr>
             @endforeach
@@ -83,8 +98,6 @@
             </tbody>
         </table>
     </div>
-    @else
-
     @endif
 @stop
 @section('scripts')
